# backend/openapi.yaml
openapi: 3.1.2
info:
  title: 于是 OOTD API
  version: 0.4.0
  description: |
    “于是”OOTD 产品的唯一后端接口契约。
    Swagger UI、iOS Swift Client 与 Web frontend 都使用此文档。
servers:
  - url: /v1
    description: Same-origin API v1
security:
  - cookieAuth: []
paths:
  /health/live:
    get:
      operationId: getLiveness
      summary: 检查 API 进程存活
      description: 仅供受限运维访问，不查询数据库，不代表云业务已启用。
      security: []
      responses:
        '200':
          description: API 进程存活
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Liveness'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /health/ready:
    get:
      operationId: getReadiness
      summary: 检查 API 接纳就绪状态
      description: 有界探测 PostgreSQL；退出或依赖故障时返回 503，不输出连接详情。
      security: []
      responses:
        '200':
          description: 当前 API 依赖可用
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Readiness'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: 暂未就绪；不能据此判断进程死亡
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Retry-After:
              schema:
                type: string
                const: '1'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/registrations:
    post:
      operationId: registerAccount
      summary: 注册账户并建立会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAccountRequest'
      responses:
        '201':
          description: 账户和会话已创建
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/EmailConflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/sessions:
    post:
      operationId: createSession
      summary: 使用邮箱密码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: 登录成功并建立新会话
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/session:
    delete:
      operationId: deleteSession
      summary: 退出当前会话
      responses:
        '204':
          description: 当前会话已失效
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me:
    get:
      operationId: getCurrentUser
      summary: 获取本人账户
      responses:
        '200':
          description: 当前登录用户
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateCurrentUser
      summary: 修改本人邮箱或显示名称
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserRequest'
      responses:
        '200':
          description: 账户已更新
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          $ref: '#/components/responses/EmailConflict'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteCurrentUser
      summary: 删除本人账户及全部会话
      responses:
        '204':
          description: 账户已硬删除
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: then_session
  headers:
    RequestID:
      description: 服务端生成的请求关联标识，不采纳客户端原始值
      schema:
        type: string
        minLength: 26
        maxLength: 64
        pattern: '^[A-Za-z0-9]+$'
    SessionCookie:
      description: HttpOnly、SameSite=Strict 会话 Cookie；生产必须包含 Secure
      schema:
        type: string
        minLength: 1
        maxLength: 512
  schemas:
    RegisterAccountRequest:
      type: object
      additionalProperties: false
      required: [email, display_name, password]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 80
        password:
          type: string
          format: password
          description: 服务端要求 12–72 个 UTF-8 字节；不按 Unicode 字符数截断
          x-minBytes: 12
          x-maxBytes: 72
      example:
        email: developer@example.test
        display_name: 开发用户
        password: example-password-2026
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        password:
          type: string
          format: password
          description: 服务端按 UTF-8 字节执行密码规则；登录失败统一返回 401
          x-maxBytes: 72
      example:
        email: developer@example.test
        password: example-password-2026
    UpdateCurrentUserRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 80
      example:
        display_name: 新显示名称
    User:
      type: object
      additionalProperties: false
      required: [id, email, display_name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 254
        display_name:
          type: string
          minLength: 1
          maxLength: 80
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 018f1f74-a2d0-7c6d-9c17-4a0ea2400a11
        email: developer@example.test
        display_name: 开发用户
        created_at: '2026-09-14T08:00:00Z'
        updated_at: '2026-09-14T08:00:00Z'
    AuthenticatedUser:
      type: object
      additionalProperties: false
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    Liveness:
      type: object
      additionalProperties: false
      required: [status, request_id]
      properties:
        status:
          type: string
          enum: [live]
        request_id:
          type: string
          minLength: 26
          maxLength: 64
          pattern: '^[A-Za-z0-9]+$'
      example:
        status: live
        request_id: TESTREQUESTIDENTIFIER00000001
    Readiness:
      type: object
      additionalProperties: false
      required: [status, request_id]
      properties:
        status:
          type: string
          enum: [ready]
        request_id:
          type: string
          minLength: 26
          maxLength: 64
          pattern: '^[A-Za-z0-9]+$'
      example:
        status: ready
        request_id: TESTREQUESTIDENTIFIER00000002
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - EMAIL_CONFLICT
            - AUTHENTICATION_FAILED
            - NOT_READY
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - INTERNAL_ERROR
        message:
          type: string
          minLength: 1
          maxLength: 160
        request_id:
          type: string
          minLength: 26
          maxLength: 64
          pattern: '^[A-Za-z0-9]+$'
        retryable:
          type: boolean
      example:
        code: AUTHENTICATION_FAILED
        message: Sign-in information is invalid.
        request_id: TESTREQUESTIDENTIFIER00000003
        retryable: false
  responses:
    BadRequest:
      description: 请求不满足接口合同
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: BAD_REQUEST
            message: Request does not satisfy the API contract.
            request_id: TESTREQUESTIDENTIFIER00000004
            retryable: false
    Unauthenticated:
      description: 会话缺失、失效或登录信息不正确
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: AUTHENTICATION_FAILED
            message: Sign-in information is invalid.
            request_id: TESTREQUESTIDENTIFIER00000005
            retryable: false
    EmailConflict:
      description: 规范化邮箱已经被其他账户使用
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: EMAIL_CONFLICT
            message: This email cannot be used.
            request_id: TESTREQUESTIDENTIFIER00000006
            retryable: false
    InternalError:
      description: 内部错误，响应不包含敏感详情
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: Internal server error.
            request_id: TESTREQUESTIDENTIFIER00000007
            retryable: false